+++ WMS/Models/Entities/WarehouseCapacityDefination.cs
@@ -6,7 +6,7 @@ namespace WMS.Models.Entities
     public partial class WarehouseCapacityDefination
     {
         public int Id { get; set; }
-        public long? WarehouseId { get; set; }
+        public int? WarehouseId { get; set; }
         public int? Row { get; set; }
         public int? Column { get; set; }
         public int? ReckQuantity { get; set; }
